+++ idf_component_tools/sources/local.py
@@ -2,6 +2,7 @@ import os
 
 from ..errors import SourceError
 from ..manifest import MANIFEST_FILENAME, ComponentWithVersions, HashedComponentVersion, ManifestManager
+from . import utils
 from .base import BaseSource
 
 try:
@@ -34,6 +35,22 @@ class LocalSource(BaseSource):
         self.source_details.get('path')
 
     def download(self, component, download_path):
+        directory_name = os.path.basename(self._path)
+        component_name_with_namespace = self.normalized_name(component.name)
+        component_name = component_name_with_namespace.replace('/', '__')
+        component_name_without_namespace = component_name.split('__')[1]
+        if component_name != directory_name and component_name.split('__')[1] != directory_name:
+            print(
+                'WARNING:  Component name "{component_name_with_namespace}" '
+                'doesn\'t match the directory name "{directory_name}". '
+                'ESP-IDF CMake build system uses directory names as names of components, '
+                'so different names may break requirements resolution. '
+                'To avoid the problem rename the component directory to '
+                '"{component_name}" or "{component_name_without_namespace}"'.format(
+                    component_name_with_namespace=component_name_with_namespace,
+                    directory_name=directory_name,
+                    component_name=component_name,
+                    component_name_without_namespace=component_name_without_namespace))
         return [self._path]
 
     def versions(self, name, details=None, spec='*', target=None):
@@ -66,3 +83,6 @@ class LocalSource(BaseSource):
             'path': self._path,
             'type': self.name,
         }
+
+    def normalized_name(self, name):
+        return utils.normalized_name(name)